move tests to PlannerHierarchySpec
